# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: real-mail-otp backend API
  version: 0.1.0
  description: >
    Temporary inbox + pay-per-read (single credit wallet, 200 VND / read) service.
    Every response carries request_id. Errors use the common envelope
    {error:{code,message,retryable}, request_id}. Cookie/key/payload/body/OTP are
    never returned in any schema.
servers:
  - url: /
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthRegisterRequest" }
      responses:
        "201":
          description: User created
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginRequest" }
      responses:
        "200":
          description: Login successful
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: Tokens refreshed
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout and revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "200":
          description: Logged out successfully
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Logged out successfully" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/billing/balance:
    get:
      tags: [billing]
      summary: Current wallet balance
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/billing/ledger:
    get:
      tags: [billing]
      summary: Wallet ledger entries
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Ledger page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/admin/payments/{id}/approve:
    post:
      tags: [admin]
      summary: Approve a payment and grant credit (admin only)
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovePaymentRequest" }
      responses:
        "200":
          description: Payment approved
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "503": { $ref: "#/components/responses/UpstreamUnavailable" }
  /v1/admin/payments/{id}/reverse:
    post:
      tags: [admin]
      summary: Reverse a paid payment (admin only)
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReversePaymentRequest" }
      responses:
        "200":
          description: Payment reversed
          headers:
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/InsufficientBalance" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
components:
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    PaymentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Not the resource owner
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Idempotency / billing conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RateLimited:
      description: Rate limited / abuse blocked
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    InsufficientBalance:
      description: Insufficient credit balance
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    UpstreamUnavailable:
      description: Upstream service unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
        request_id: { type: string }
...
    Balance:
      type: object
      required: [balance_vnd]
      properties:
        balance_vnd: { type: integer, example: 20000 }
    LedgerEntry:
      type: object
      required: [id, type, amount_vnd, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [credit, debit, reversal] }
        amount_vnd: { type: integer }
        reference_type: { type: string }
        reference_id: { type: string }
        created_at: { type: string, format: date-time }
    LedgerList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LedgerEntry" }
        next_cursor: { type: [string, "null"] }
...
    Payment:
      type: object
      required: [id, amount_vnd, status, created_at]
      properties:
        id: { type: string, format: uuid }
        package_code: { type: [string, "null"] }
        amount_vnd: { type: integer }
        provider_ref: { type: string }
        qr_content: { type: string }
        status:
          type: string
          enum: [pending, pending_review, paid, rejected, expired]
        created_at: { type: string, format: date-time }
        paid_at: { type: [string, "null"], format: date-time }

    # --- Auth schemas (P1-02) ---
    AuthRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }

    # --- Admin schemas (P0-02, P0-04) ---
    ApprovePaymentRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 500 }
    ReversePaymentRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 500 }